Refatorando Header e reorganizando pastas com Pages e components
