+++ src/App.js
@@ -2,8 +2,9 @@ import React from 'react';
 import { BrowserRouter, Route, Switch } from 'react-router-dom';
 import './App.css';
 
-import Home from './components/Home';
-import Game from './components/Game';
+import Home from './Pages/Home';
+import Game from './Pages/Game';
+import Feedback from './Pages/Feedback';
 import Settings from './components/Settings';
 
 export default function App() {
@@ -13,6 +14,7 @@ export default function App() {
         <Route exact path="/" component={Home} />
         <Route path="/gameplay" component={Game} />
         <Route path="/settings" component={Settings} />
+        <Route path="/feedback" component={Feedback} />
       </Switch>
     </BrowserRouter>
   );
