+++ src/Pages/Feedback.jsx
@@ -1,0 +1,15 @@
+import React from 'react';
+import Header from '../components/Header';
+
+class Feedback extends React.Component {
+  render() {
+    return (
+      <div className="container-header">
+        <header>
+          <Header />
+        </header>
+      </div>
+    );
+  }
+}
+export default Feedback;
